fix(qdmi): correct T2 decoherence bound and shots type in adapter
The T1 >= T2 assertion violated quantum physics — the correct bound is
T2 <= 2*T1 (from 1/T2 = 1/(2T1) + 1/T_phi). Also fix shots parameter
in system-qdmi adapter path: was sending c_int (4 bytes) instead of
size_t (8 bytes) per QDMI v1.2.1 spec.

+++ adapters/arvak-adapter-qdmi/src/backend.rs
@@ -681,13 +681,13 @@ impl Backend for QdmiBackend {
                 ffi::check_status(status)
                     .map_err(|s| HalError::Backend(format!("set Program failed: {s:?}")))?;
 
-                // 4. Set shots
-                let shots_val = shots as c_int;
+                // 4. Set shots (QDMI expects size_t / usize)
+                let shots_val = shots as usize;
                 let status = ffi::QDMI_job_set_parameter(
                     job,
                     QdmiJobParameter::ShotsNum as c_int,
-                    std::mem::size_of::<c_int>(),
-                    &shots_val as *const c_int as *const c_void,
+                    std::mem::size_of::<usize>(),
+                    &shots_val as *const usize as *const c_void,
                 );
                 ffi::check_status(status)
                     .map_err(|s| HalError::Backend(format!("set ShotsNum failed: {s:?}")))?;
